+++ alpha/web/index.php
@@ -219,11 +219,14 @@ function checkCache()
 	else if (strpos($uri, "/serveFlavor/") !== false && function_exists('apc_fetch') && $_SERVER["REQUEST_METHOD"] == "GET")
 	{
 		require_once(dirname(__FILE__) . '/../apps/kaltura/lib/renderers/kRendererDumpFile.php');
+		require_once(dirname(__FILE__) . '/../apps/kaltura/lib/renderers/kRendererString.php');
 		require_once(dirname(__FILE__) . '/../apps/kaltura/lib/monitor/KalturaMonitorClient.php');
+		require_once(dirname(__FILE__) . '/../apps/kaltura/lib/request/kIpAddressUtils.php');
 		
 		$host = isset($_SERVER['HTTP_X_FORWARDED_HOST']) ? $_SERVER['HTTP_X_FORWARDED_HOST'] : $_SERVER['HTTP_HOST'];
+		$cacheKey = 'dumpFile-'.kIpAddressUtils::isInternalIp().'-'.$host.$uri;
 		
-		$renderer = apc_fetch("dumpFile-{$host}{$uri}");
+		$renderer = apc_fetch($cacheKey);
 		if ($renderer)
 		{
 			KalturaMonitorClient::initApiMonitor(true, 'extwidget.serveFlavor', $renderer->partnerId);